+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # AEMO Gem Changelog
 
+## 0.3.1
+
+* Catch invalid length  NEM12 records (300 and 400)
+
 ## 0.3.0
 
 * Refactor NMI allocations